requirements.txt
alembic==1.4.3
aniso8601==8.0.0
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
Werkzeug==1.0.1
